refactor: Info 컴포넌트 common에서 export
[#8]
