+++ packages/frontend/src/design-system/components/common/Info/index.tsx
@@ -1,0 +1,3 @@
+import Info from "./Info";
+
+export default Info;
