+++ packages/frontend/src/design-system/components/common/index.ts
@@ -6,3 +6,4 @@ export { Badge, badgeVariantList } from "./Badge";
 export { toast, ToastContainer } from "./Toast";
 export { Accordion, useAccordion } from "./Accordion";
 export { default as Footer } from "./Footer";
+export { default as Info } from "./Info";
